Removing a bunch of auto-generated files from deploy directory
Updated "make generate" so that it only applies to CRDs YAML files
and golang code that is autogenerated from pkg/api source code

Moved YAML install file generate to the build/package.sh script.
This can now be used to generate a release-<VERSION> directory
with all the relevant files for a new release of Splunk Operator.