Updated build/make_bundle.sh to also create a red hat certified bundle in the deploy/olm-certified directory
Added finalizers to examples in CSV files so that the PVCs will get
cleaned up with the CR instances.

Flipped AllNamespaces support to false. This seems to mostly work
except security control contexts need to be manually added to any
namespaces where you want to use it. Decided it's better to just
leave it off to prevent confusion, until we can figure out a better
approach for that..